refactor: consolidate DI adapter implementation files
- Merge functionality from di_adapter.py into fastapi_injectable_adapter.py
- Add scope conversion and get_service_by_type convenience function
- Update import references in dependent files
- Remove redundant di_adapter.py
